Added [33m       _  _ __ | | __ _ _  _ | '_ \| |/ _' | || | |  __/|_|\____|\__ / |_|            |__/
[0mplay 2.2.3 built with Scala 2.10.3 (running Java 1.7.0_07), http://www.playframework.com
[31m
This is not a play application!
[0m
Use `play new` to create a new Play application in the current directory,
or go to an existing application and launch the development console using `play`.

You can also browse the complete documentation at http://www.playframework.com. command during webapp run in order to re-package changes in cfp-src